Рефакторинг кода

+++ hardware/src/ru/mydiy/hw/MotionSensorListener.java
@@ -3,6 +3,6 @@ package ru.mydiy.hw;
 public interface MotionSensorListener {
     void motionState(MotionSensor monitor, boolean state);
     void onException(MotionSensor monitor, Exception e);
-    void activityIsGone(MotionSensor monitor, int overallTime);
+    void activityIsGone(MotionSensor monitor, long overallTime);
     void debugMessage(MotionSensor monitor, String msg);
 }
